requirements.txt
fastapi
uvicorn[standard]
crewai
openai
requests
python-dotenv
pydantic
